move into src/sentry/seer
